new signature actor. remove content checker
