+++ kifi-backend/modules/shoebox/app/com/keepit/scraper/ScraperPlugin.scala
@@ -23,6 +23,7 @@ import com.keepit.inject.AppScoped
 
 case object Scrape
 case class ScrapeInstance(uri: NormalizedURI)
+case class GetSignature(url: String)
 
 private[scraper] class ScraperActor @Inject() (
     scraper: Scraper,
@@ -38,13 +39,25 @@ private[scraper] class ScraperActor @Inject() (
   }
 }
 
+private[scraper] class ScrapeSignatureActor @Inject() (
+  scraper: Scraper,
+  healthcheckPlugin: HealthcheckPlugin
+) extends FortyTwoActor(healthcheckPlugin) with Logging {
+  def receive() = {
+    case GetSignature(url) => sender ! scraper.getSignature(url)
+    case m => throw new Exception("unknown message %s".format(m))
+  }
+}
+
 trait ScraperPlugin extends Plugin {
   def scrapePending(): Future[Seq[(NormalizedURI, Option[Article])]]
   def asyncScrape(uri: NormalizedURI): Future[(NormalizedURI, Option[Article])]
+  def asyncSignature(url: String): Future[Option[Signature]]
 }
 
 class ScraperPluginImpl @Inject() (
     actorProvider: ActorProvider[ScraperActor],
+    signatureActorProvider: ActorProvider[ScrapeSignatureActor],
     scraper: Scraper,
     val schedulingProperties: SchedulingProperties) //only on leader
   extends ScraperPlugin with SchedulingPlugin with Logging {
@@ -65,6 +78,9 @@ class ScraperPluginImpl @Inject() (
   override def scrapePending(): Future[Seq[(NormalizedURI, Option[Article])]] =
     actorProvider.actor.ask(Scrape)(1 minutes).mapTo[Seq[(NormalizedURI, Option[Article])]]
 
-  override def asyncScrape(uri: NormalizedURI): Future[(NormalizedURI, Option[Article])] = 
+  override def asyncScrape(uri: NormalizedURI): Future[(NormalizedURI, Option[Article])] =
     actorProvider.actor.ask(ScrapeInstance(uri))(1 minutes).mapTo[(NormalizedURI, Option[Article])]
+
+  override def asyncSignature(url: String): Future[Option[Signature]] =
+    signatureActorProvider.actor.ask(GetSignature(url))(1 minutes).mapTo[Option[Signature]]
 }
